Update the volume ID substitutions list and application
Some newer subvariants / loadouts, and 'Rawhide' for some
reason, were missing from the substitution list. This updates
it, alphabetizes it, adjusts the indentation to match how it
is in pungi-fedora for ease of comparison, and also tweaks the
way the list is applied to match
https://pagure.io/pungi/pull-request/857 , the correct way to
ensure longer substitutions are applied before shorter ones.

Note 'Alpha' and 'TC' are removed because we don't do Alphas or
TCs any more.

Signed-off-by: Adam Williamson <awilliam@redhat.com>