hub: unlimited NameWidth for kojifiles Apache location
Prior to this change, Apache's mod_autoindex would limit the file name
lengths in when it displayed directory and file lists under the archive
location /kojifiles.

Koji archive filenames are often be longer than mod_autoindex's default
of 20 bytes.

Configure mod_autoindex to show the entire filenames. This makes it
easier for users to browse /kojifiles to see at a glance which files are
stored.