builder/kojid: revert using only 2 processors for xz compression and lower preset to 6
In c360c791ec24b567a16553dca38b24645dd80a36 we changed xz to use only
2 cpus for compression. That commit notes that xz threads are really
memory hungry. However, preset -9 is many times as memory hungry as
preset -6. If we switch to preset 6, xz compressed artifacts will be
a little bit larger, but will compress (and decompress!) a lot larger.
This is obviously a trade off, but I think lowering the preset and
using more cpus will be a better tradeoff.

This commit moves xz to preset 6 and using all cpus the builder has.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>