hub: Fix issue with listing users and old versions of Postgres
Koji was using the array_remove() SQL function, which was not introduced
until Postgres 9.4.  RHEL 7 ships with Postgres 9.2.  This construct
was being used to ensure that the krb_principals arrays in user
dictionaries were empty if no principals were defined, instead of
having a single value of None.  Instead we use QueryProcessor's result
transformation support to remove the None entry.

Signed-off-by: Jim Foraker <foraker1@llnl.gov>