Disable notifications from clone-tag by default
The tagBuildBypass and untagBuildBypass invocations in the clone-tag
CLI were not setting the notify option, leaving it at the default of
True. This means in some cases many hundreds of notification events
would be triggered from a clone-tag call.

Change the default behavior of clone-tag to explicitly disable
notifications

Add a new --notify option to clone-tag to re-enable notifications for
tagging and untagging of builds.

Fixes: #1890