hub: simplify recipients condition in build_notification()
If get_notification_recipients() returns a non-empty list of recipients,
then the build_notification() method creates a new buildNotification task.

Simplify the conditional that checks the get_notification_recipients()
return value so that it is easier to read. Python treats an empty list
as False and a non-empty list as True, so we don't need to explicitly
test the list length here.

The purpose of this change is to make this code easier to read and
understand.