Fix PR#2347 - make "not wait" work again
Prior to PR#2347, koji would not wait for image
and indirectionimage builds even if --wait was
specified. After PR#2347, koji always waits,
even when --wait is NOT specified.

According to the CLI docs, "wait" should happen
regardless of whether we've decided to run in
the background, thus, checking this is superfluous.

Signed-off-by: Lon Hohberger <lhh@redhat.com>