cli: use main requests.Session in download_file()
Prior to this change, we used new requests.Session objects in the
download_file() method. This meant that we did not respect the
"serverca" profile setting (for example), and we would inefficiently
recreate a new HTTP session for every file we downloaded.

Update the CLI code to pass the main Koji session's rsession (requests
Session) through to download_file(), so we can re-use our HTTP settings
and connection.