config file on the command line should replace the system config
Koji config files are additive.  That makes sense when loading
the system config, followed by /etc/koji.d/*. But when a user
config is provided on the command line, it should be loaded
instead of the system config and not in addition to it. Otherwise
the system config can never be fully overridden.