kiwi: Bind-mount /dev into the buildroot for old-chroot
The OEM images can't be build with nspawn as they require mocking with
device-mapper which is near to impossible to properly to set up in the
container so those ones have to be built in old chroot where /dev can
and must be bind-mounted. In opposite for nspawn, bind-mounting /dev
breaks build completely.

Let's just dynamically decide based on the chroot type what to do.