kiwi: rename 'vhdfixed' images to 'vhd'
Kiwi's "vhd-fixed" format, which we use to generate Azure images,
outputs images with the extension "vhdfixed.xz". This isn't in
Koji's list of accepted image extensions, and updating that
involves a database schema update, which is a drag. It seems
easier to just rename these images to "vhd.xz" when uploading
them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>